[netutil] Improve macOS support
