+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `[fmtc]` Added TrueColor support
 * `[system]` Improved macOS support
+* `[netutil]` Improved macOS support
 * `[system]` Fixed bug with extracting macOS version
 * `[system]` Fixed bug with extracting macOS arch info
 * `[fmtc]` Code refactoring
